# ENH: Speed up next_fast_len with LRU cache