On `master` the new benchmark gives:
```
                 size               
              --------- ------------
                  12      592±7ns   
                  13      672±70ns  
                 1021     821±40ns  
                 1024     729±70ns  
                16381    1.05±0.3μs 
                16384     678±40ns  
                262139   1.61±0.1μs 
                262144    716±40ns  
                999983   1.74±0.2μs 
               1048576    723±30ns
```
And on this PR, implementing @pv's alternative idea from #10589:
```
                 size             
              --------- ----------
                  12     196±10ns 
                  13     205±10ns 
                 1021    208±20ns 
                 1024    193±30ns 
                16381    196±10ns 
                16384    201±30ns 
                262139   172±10ns 
                262144   211±20ns 
                999983   195±10ns 
               1048576   186±20ns 
```
LRU default cache size is 128 which seems reasonable for storing these.