@larsoner 2019-09-10
@toddrjen given the speed I marked this as closing #10589, but let me know if you see that it's insufficient in some local benchmarks for your use case.